requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
opencv-python-headless
numpy
pandas
matplotlib
pillow
tqdm
python-dotenv  
langchain
langchain-groq
langchain-community
langchain-core
langchain_huggingface
langchain_chroma
onnxruntime
requests
pydantic
pypdf
jinja2